requirements.txt
opencv-python
time-python
imagegrab-python
pymysql-python
paho.mqtt.client-python
RPi.GPIO-python
spidev-python
